chore(tests): Small refactor of file upload for nrs
